# specs/006-domain-class-api-contracts/contracts/openapi/student-registration-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Student Registration API",
    "version": "v1"
  },
  "paths": {
...
    "/api/auth/student/login": {
      "post": {
        "tags": [
          "Spec007Endpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/student/activate": {
      "post": {
        "tags": [
          "Spec007Endpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateStudentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/staff/login": {
      "post": {
        "tags": [
          "Spec007Endpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Spec007Endpoints"
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "StudentRegistration.Identity": []
          }
        ]
      }
    },
    "/api/auth/recovery/request": {
      "post": {
        "tags": [
          "Spec007Endpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
    "/api/auth/recovery/complete": {
      "post": {
        "tags": [
          "Spec007Endpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/password/change": {
      "post": {
        "tags": [
          "Spec007Endpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "StudentRegistration.Identity": []
          }
        ]
      }
    },
    "/api/auth/sessions/revoke-all": {
      "post": {
        "tags": [
          "Spec007Endpoints"
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "StudentRegistration.Identity": []
          }
        ]
      }
    },
    "/api/auth/session": {
      "get": {
        "tags": [
          "Spec007Endpoints"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "StudentRegistration.Identity": []
          }
        ]
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": [
          "Spec007Endpoints"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfIdentityUserSummaryDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "StudentRegistration.Identity": []
          }
        ]
      }
    },
    "/api/admin/users/imports": {
      "post": {
        "tags": [
          "Spec007Endpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityImportBatchDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "StudentRegistration.Identity": []
          }
        ]
      }
    },
    "/api/admin/users/imports/{importId}": {
      "get": {
        "tags": [
          "Spec007Endpoints"
        ],
        "parameters": [
          {
            "name": "importId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityImportBatchDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "StudentRegistration.Identity": []
          }
        ]
      }
    },
    "/api/admin/users/imports/{importId}/publish": {
      "post": {
        "tags": [
          "Spec007Endpoints"
        ],
        "parameters": [
          {
            "name": "importId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityImportPublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityImportBatchDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "StudentRegistration.Identity": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivateStudentRequest": {
        "required": [
          "universityId",
          "initialPassword",
          "newPassword"
        ],
        "type": "object",
        "properties": {
          "universityId": {
            "type": "string"
          },
          "initialPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        }
      },
...
      "ApiError": {
        "required": [
          "code",
          "message",
          "correlationId"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "currentVersion": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ChangePasswordRequest": {
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        }
      },
...
      "RecoveryCompleteRequest": {
        "required": [
          "challengeToken",
          "newPassword"
        ],
        "type": "object",
        "properties": {
          "challengeToken": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        }
      },
      "RecoveryRequest": {
        "required": [
          "universityIdOrUserName"
        ],
        "type": "object",
        "properties": {
          "universityIdOrUserName": {
            "type": "string"
          }
        }
      },
...
      "SessionDto": {
        "required": [
          "displayName",
          "roles",
          "activeRole",
          "sessionState",
          "expiresAtUtc"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "activeRole": {
            "type": [
              "null",
              "string"
            ]
          },
          "sessionState": {
            "type": "string"
          },
          "expiresAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "StaffLoginRequest": {
        "required": [
          "userName",
          "password"
        ],
        "type": "object",
        "properties": {
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "StudentLoginRequest": {
        "required": [
          "universityId",
          "password"
        ],
        "type": "object",
        "properties": {
          "universityId": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "StudentRegistration.Identity": {
        "type": "apiKey",
        "name": "__Host-StudentRegistration.Session",
        "in": "cookie"
      }
    }
  },
...
}